Moved polling station e2e tests into polling-station folder
